Remove first step, fix translation and loading
