+++ packages/rocketchat-i18n/i18n/en.i18n.json
@@ -3852,7 +3852,7 @@
   "Teams_convert_channel_to_team": "Convert to Team",
   "Teams_delete_team_choose_channels": "Select the Channels you would like to delete. The ones you decide to keep, will be available on your workspace.",
   "Teams_delete_team_public_notice": "Notice that public Channels will still be public and visible to everyone.",
-  "Teams_delete_team_warning": "This can’t be undone. Once you delete a Team, all chat content and configuration will be deleted.\n\n In the next step, you can choose if you will keep or delete Channels.",
+  "Teams_delete_team_Warning": "Once you delete a Team, all chat content and configuration will be deleted.",
   "Teams_delete_team": "You are about to delete this Team.",
   "Teams_deleted_channels": "The following Channels are going to be deleted:",
   "Teams_Errors_Already_exists": "The team `__name__` already exists.",
